Update README.md
Here we use a new method to study the correlation between different molecular emission lines in star-forming molecular dust clouds. This method is applied to maps from the Line Emission as a Tool for Galaxy Observations project in Orion-A. Our data shows the correlation properties change within dust clouds in Orion-A, suggesting that the physical properties change as well. We present our code and data in an accessible online repository.